requirements.txt
Django==1.10
djangorestframework==3.8.2
pkg-resources==0.0.0
